+++ bin/examples/repl.dart
@@ -1,6 +1,10 @@
 import 'dart:convert';
 import 'package:puredartlua/runner.dart';
 
+/// Custom error handlers can act on reported problems however you want.
+/// In this example, we just print to conslole via [print].
+void handleErrors(List<String> errs) => errs.forEach(print);
+
 /// This example driver demonstrates a very simple REPL interpreter.
 /// Read Evaluate Print Loop.
 void main() {
@@ -23,6 +27,10 @@ void main() {
       content += input;
     }
 
+    // Two special commands take priority:
+    // If 'exit' => quit REPL
+    // If 'dump_scope' => print the current scope to console.
+    // Else => evaluate string as lua script.
     switch (content.toLowerCase()) {
       case 'exit':
         loop = false;
@@ -33,15 +41,22 @@ void main() {
         continue;
     }
 
-    final ast = parse(content, onErrors: (errs) => errs.forEach(print));
+    // Try to parse and consume the input we built.
+    final ast = parse(content, onErrors: handleErrors);
     content = '';
+
+    // Errors will already be collected. We cannot execute without a valid AST.
     if (ast == null) continue;
 
-    final (ok, out) = runner(
+    // Run with a custom evaluator that clears previous results each
+    // time the constructor is called.
+    final (_, out) = runner(
       ast,
       constructor: () => evaluator..clearResults(),
-      onErrors: (errs) => errs.forEach(print),
+      onErrors: handleErrors,
     );
+
+    // Display to user in console.
     print(out);
   }
 }